refactor: suppress dead_code warnings and optimize regex compilation
- Add #[allow(dead_code)] attributes to unused struct fields across search engine modules (ads, braveapi, core_engine, discourse, duckduckgo_definitions, google_images, invidious)
- Compile regex pattern once outside loop in GoogleScholar to improve performance
- Remove unused SearchCategory imports from google_news and google_videos
- Fix type casting in MariaDbEngine offset calculation for consistency
- Suppress compiler warnings for intentionally unused deserialized fields that may be needed for future functionality or API compatibility

+++ crates/metasearch-engine/src/mariadb_engine.rs
@@ -112,7 +112,7 @@ impl SearchEngine for MariaDbEngine {
         {
             use mysql_async::{prelude::*, Pool, Opts, OptsBuilder};
 
-            let offset = (query.page.saturating_sub(1)) * self.limit;
+            let offset = (query.page.saturating_sub(1)) * (self.limit as u32);
             let wildcard = format!("%{}%", query.query.replace(' ', "%"));
 
             let opts: Opts = OptsBuilder::default()
